pyproject.toml
[project]
name = "anonymize-me"
version = "0.1.0"
description = "Streamlit application for anonymizing PI/PII data in email content and attachments."
authors = [{ name = "Anonymize Me" }]
requires-python = ">=3.11"
dependencies = [
    "streamlit",
    "openai>=1.0.0",
    "PyMuPDF>=1.23.0",
    "python-docx>=1.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]
